netutils/libcoap : minor patch to fix build warning on CPP compiler
This commit is minor patch to fix build warning on CPP compiler
1. compile warning log
warning: comparison between signed and unsigned integer expressions
[-Wsign-compare]
  if (size <= (bit >> 3)) {
2. solution
adding casting operator after bit shift operations

+++ apps/include/netutils/libcoap/bits.h
@@ -48,7 +48,7 @@
  */
 inline static int bits_setb(uint8_t *vec, size_t size, uint8_t bit)
 {
-	if (size <= (bit >> 3)) {
+	if (size <= (size_t)(bit >> 3)) {
 		return -1;
 	}
 
@@ -69,7 +69,7 @@ inline static int bits_setb(uint8_t *vec, size_t size, uint8_t bit)
  */
 inline static int bits_clrb(uint8_t *vec, size_t size, uint8_t bit)
 {
-	if (size <= (bit >> 3)) {
+	if (size <= (size_t)(bit >> 3)) {
 		return -1;
 	}
 
@@ -89,7 +89,7 @@ inline static int bits_clrb(uint8_t *vec, size_t size, uint8_t bit)
  */
 inline static int bits_getb(const uint8_t *vec, size_t size, uint8_t bit)
 {
-	if (size <= (bit >> 3)) {
+	if (size <= (size_t)(bit >> 3)) {
 		return -1;
 	}
